Wire shadow systems into render pipeline and fix build warnings
- Wire PostProcessingPipeline, TemporalEffects, ShadowAtlas, and
  ScreenSpaceEffects into GraphicsEngine lifecycle (init, render,
  resize, shutdown) — these were built but never called
- Add ShadowAtlas: priority-based shadow map tile allocator with
  grid subdivision, eviction, and metrics
- Implement full frustum culling on Linux/RHI path using array-indexed
  matrix (cross-platform alternative to Windows named members)
- Fix compiler warnings: printf format specifiers (GameObject, NetworkManager),
  unused variable (SceneSnapshotSerializer), always-true comparison
  (TestLocalizationSystem), enum class streaming (TestHybridRT)
- Add TestShadowAtlas with 7 unit tests for tile allocation system

+++ Tests/TestHybridRT.cpp
@@ -32,7 +32,7 @@ TEST(HybridRT_QualityEnum_HasAllValues)
 TEST(HybridRT_RTCapabilities_DefaultDisabled)
 {
     RTCapabilities caps;
-    EXPECT_EQ(caps.bestBackend, RayTracingBackend::Disabled);
+    EXPECT_EQ(static_cast<int>(caps.bestBackend), static_cast<int>(RayTracingBackend::Disabled));
     EXPECT_FALSE(caps.supportsHardwareRT);
     EXPECT_FALSE(caps.supportsInlineRT);
     EXPECT_EQ(caps.maxRecursionDepth, 0u);
@@ -45,7 +45,8 @@ TEST(HybridRT_RTCapabilities_InDeviceCaps)
     // Verify RTCapabilities is accessible as a member
     deviceCaps.rayTracing.bestBackend = RayTracingBackend::Software_SDFGI;
     deviceCaps.rayTracing.supportsHardwareRT = false;
-    EXPECT_EQ(deviceCaps.rayTracing.bestBackend, RayTracingBackend::Software_SDFGI);
+    EXPECT_EQ(static_cast<int>(deviceCaps.rayTracing.bestBackend),
+              static_cast<int>(RayTracingBackend::Software_SDFGI));
     EXPECT_FALSE(deviceCaps.rayTracing.supportsHardwareRT);
 }
